move FuelInfo to translator utils from func submodule
